V0.1.1 - SaveNewRace action completed in HomeController

+++ Models/Race.cs
@@ -12,7 +12,7 @@ namespace RagnarEstimator.Models
     public DateTime RaceEnd {get; set;}
     public DateTime Sunset {get; set;}
     public DateTime Sunrise {get; set;}
-    public float RacePaceMultiplyer {get; set;}
+    public decimal RacePaceMultiplyer {get; set;}
     public string RaceName {get; set;}
     public bool Type {get; set;}
     public string RaceNotes {get; set;}
